Search field on the top of the search screen is replaced with SearchView for Android >= 3.0 (API 11).
dev notes:
Api11_compat class is defined as inner class of Search2Activity. The reason is that we want to prevent the dalvik verifier to be allergic to references to SearchView when loading Search2Activity class.
By defining Api11_compat as inner class, it's considered as a separate class when compiled, and we make it such that Api11_compat is not instantiated at all on runtime when we do not use an API 11 device, so dalvik verifier won't complain anything.