Reduce activity task stack height when starting main activities from the left drawer.
Task stack will be reduced to just the Bible text when "Bible" menu item from the left drawer is pressed.
Other menu items "Devotion", "Songs", "Reading Plan" when pressed, will cause task to be cleared, then Bible text activity to be started in the background, and the corresponding activity in the foreground.