Add sync settings screen. Configure GCM. Sync on start up and on receiving GCM message.
On the settings screen user can log in to a Google account. ID Token will be obtained and sent to the sync server.

Registration id for GCM will be obtained on app startup, if the stored one does not exist or exists but corresponds to a different app versionCode.
When a new GCM registration id is obtained, it is sent to the server. This also happens when the user logs in to an account on the sync settings screen.