Handle multiple highlights in a verse.
When there are multiple highlights in a verse (this can only happen logically via sync), show the color of the
most recent one (based on modifyTime).

If the highlight color is modified (or removed), edit (or delete) the most recent row in the db table, but also remove the old rows if they exist.